Introduce ProductTagLookupParams to encapsulate the growing parameters
We keep finding ourselves adding parameters to the SubscriptionDefinition.getAllProductTags and it's tedious to maintain.

Introduced ProductTagLookupParams to encapsulate these.  When we need to add another parameter, we can do it here without modifying the signature of the method.

Then with the new predicate-building approach in that method, it will be easy enough to expand on with something like

```java
Predicate<Variant> params.newParameter == null ? variant -> true : variant -> whatever == params.getNewParameter()
```